New Python SPEFit function and Script
