+++ python/SPEFit.py
@@ -67,8 +67,12 @@ def Fit_2_gauss(h,hped,HV,UseHped=False,RobustMode=False,FreeMode = False, Light
     index = 0
     while (PositiveTable == 0):
         PositiveTableTemp = np.where( diffTable[MaxIndex+index:] > 0 )
-        PositiveTable = PositiveTableTemp[0][0]
-        index = index+1
+        if PositiveTableTemp:
+            PositiveTable = PositiveTableTemp[0][0]
+            index = index+1
+        else:
+            PositiveTable = len(histCharge)-1
+			
     # ~ print(PositiveTable)
     # ~ print("here")
     # ~ print(histTable[:MaxIndex+PositiveTable])
@@ -187,6 +191,8 @@ def Fit_1_gauss_bis(h,hped,iv,Optimizer=False,Verbose=False,WallTime=2,FixedPara
     ses_2g = calin.math.pdf_1d.BinnedGaussianPDF(h.dxval())
     ses_2g.set_parameter_values(np.asarray([1,0.5]))
     #mes_2g = calin.calib.spe_fit.GeneralPoissonMES(-200, h.dxval(), 4094, ses_2g, ped) #test benche
+    global MinCharge
+    global NSample
     mes_2g = calin.calib.spe_fit.GeneralPoissonMES(MinCharge, h.dxval(), NSample, ses_2g, ped) #test bench
     
     if (UseHped):
@@ -297,8 +303,8 @@ def Fit_2_gauss_bis(h,hped,iv,Optimizer=False,Verbose=False,WallTime=2,FixedPara
                 limitUp[i] = iv2[i] + np.sign(iv2[i])*iv2[i]*0.08
                 Stepsize[i] = abs(iv2[i])*0.01
             if (i == 1):
-                limitLow[i] = iv2[i] - np.sign(iv2[i])*iv2[i]*0.3
-                limitUp[i] = iv2[i] + np.sign(iv2[i])*iv2[i]*0.3
+                limitLow[i] = iv2[i] - np.sign(iv2[i])*iv2[i]*0.15
+                limitUp[i] = iv2[i] + np.sign(iv2[i])*iv2[i]*0.15
                 Stepsize[i] = abs(iv2[i])*0.01  
             if ((i in FixedParam) and i < 3):
                 limitLow[i] = iv2[i] - np.sign(iv2[i])*iv2[i]*1e-15
@@ -347,89 +353,3 @@ def Fit_2_gauss_bis(h,hped,iv,Optimizer=False,Verbose=False,WallTime=2,FixedPara
             plot(xhist,ymodel_2g, 'r', label='Gaussian')
         a=list(axis())
     return (xopt_2g,xerr,mes_2g,limitLow,limitUp,status,ses_2g,ped)
-
-
-def runScript():
-	DataListSignal = []
-	DataListPedestal = []
-	#Filename = '_OD1_18_16_dark_opt_ASCII_histo'
-	
-	# Filename is a launch argument
-	Filename = sys.argv[1]
-
-	#DataListSignal.append(calin.iact_data.llr.make_lecroy_adc_hist(sys.argv[1],scale=1))
-	# ~ DataListSignal = calin.iact_data.llr.make_lecroy_adc_hist('/CTA/DataTestBench/PierreJeanTestBench/spectrum_spe_ch0_1540V.txt',scale=1)
-	# ~ DataListSignal = calin.iact_data.llr.make_lecroy_adc_hist('/CTA/SPE_fitter/data/_OD0_flash7.8V_LED7_0.3mA_ASCII_histo.txt', scale=1)
-    
-    # Loading of the file and contruction of the calin histogram called DataListSignal
-    # here delimiter are , but can be change to other stuffs, skiprows is the number of column that we have to skip at the begin of the ASCII file
-	scale=1
-	data = np.loadtxt('/CTA/DataTestBench/NewSetup/'+Filename+'.txt', delimiter=',', skiprows=5)
-	dx = np.median(data[1:,0] - data[0:-1,0])/scale
-	DataListSignal = calin.math.histogram.SimpleHist(dx)
-	DataListSignal.insert_two_vec(data[:,0]/scale, data[:,1])
-	
-
-	SimpleModel = False
-
-	if (not SimpleModel):
-		# 2 Gauss Fit with constrained model, fast mode
-		# Fit_2_gauss(hsignal,hped,HV), hped is not used except if you specify Fit_2_gauss(hsignal,hped,HV,True) 
-		# On the following I put hped = DataListSignal, because hped is not used
-		# output : result is a table with parameters
-		# aerror is a table with their uncertainty
-		
-		result,aerror,b,c,d = Fit_2_gauss(DataListSignal,DataListSignal,1000)
-		
-		# 2 Gauss Fit with constrained model, Robust mode, all parameters free, should only work at high voltage (HV > 1200)
-		# Use it at your own risk
-		# ~ result,aerror,b,c,d = Fit_2_gauss(DataListSignal,DataListSignal,1000,False,True,True)
-	else:
-		result,aerror,b,c,d = Fit_1_gauss(DataListSignal,DataListSignal,1000)
-
-	print(aerror)
-	mes_2g = b
-	
-    # plot the result on a pdf called "myfig"
-    
-	calin.plotting.plot_histogram(DataListSignal, lw=2, label='SPE data')
-	plt.xlabel('Signal [DC]')
-	plt.ylabel('Events per %d DC bin'%DataListSignal.dxval())
-
-	ihist = range(0,DataListSignal.nbin())
-	xhist = DataListSignal.all_xval_center()
-
-	mes_2g.set_parameter_values(result)
-	ymodel_2g = \
-		list(map(lambda x: DataListSignal.sum_w()*DataListSignal.dxval()*mes_2g.pdf_mes(x),xhist))
-	matplotlib.pyplot.plot(xhist,ymodel_2g, 'r', label='Two Gaussian')
-	a=list(plt.axis())
-	plt.axis(a)
-	plt.legend()
-	plt.grid()
-	plt.savefig('/CTA/myfig')
-
-    # Compute the mean of the single electron spectrum, this is the gain
-    #Compute also the variance and the resolution
-    
-	ses_mean = sum(mes_2g.all_ses_x() * mes_2g.single_electron_spectrum())*DataListSignal.dxval()
-	ses_var = sum(mes_2g.all_ses_x()**2 * mes_2g.single_electron_spectrum())*DataListSignal.dxval()-ses_mean**2
-	ses_res = np.sqrt(ses_var)/ses_mean
-
-	ses_mean_uncertainty = 0
-	if (not SimpleModel):
-		# Compute the uncertainty of the gain using the uncertainty of the parameters high gauss mean and resolution
-		ses_mean_uncertainty = GainUncertainty(0.715,0.45,result[3],result[4],aerror[4],aerror[3])
-		print("Gain Uncertainty")
-		print(ses_mean_uncertainty)
-
-	F = open("/CTA/SPE_fitter/"+Filename+"Output.txt","w")
-	print("SPE: spe_mean = %-6.2f ; spe_var = %-6.2f; spe_res = %-6.3f; "%(ses_mean,np.sqrt(ses_var),np.sqrt(ses_var)/ses_mean))
-	for i in range(len(a)):
-		F.write(str(result[i])+'\t')
-		F.write(str(aerror[i])+'\n')
-	F.write(str(ses_mean)+'\t')
-	F.write(str(ses_mean_uncertainty)+'\n')
-	F.close()
-
-runScript()
